fix cloud service and remve strtobool of coud utils tooo
