+++ cloud-service-api/utils/utils.py
@@ -6,22 +6,6 @@ from flask import current_app as app
 from models.model import Trabajador
 
 
-# https://stackoverflow.com/questions/42248342/yes-no-prompt-in-python3-using-strtobool
-def strtobool(val):
-    """Convert a string representation of truth to true (1) or false (0).
-    True values are 'y', 'yes', 't', 'true', 'on', and '1'; false values
-    are 'n', 'no', 'f', 'false', 'off', and '0'.  Raises ValueError if
-    'val' is anything else.
-    """
-    val = val.lower()
-    if val in ('y', 'yes', 't', 'true', 'on', '1', 'activa', 'activo'):
-        return True
-    elif val in ('n', 'no', 'f', 'false', 'off', '0', 'desactiva', 'inactiva', 'inactivo'):
-        return False
-    else:
-        raise ValueError("invalid truth value %r" % (val,))
-
-
 # https://www.bacancytechnology.com/blog/flask-jwt-authentication
 def token_required(f):
     @wraps(f)
